+++ src/Strand.php
@@ -1,0 +1,4 @@
+<?php
+namespace Icicle\Concurrent;
+
+interface Strand extends Context, Sync\Channel {}